도메인은 결과정보만 저장하도록 변경
