+++ src/main/resources/application.properties
@@ -21,3 +21,5 @@ logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
 handlebars.suffix=.html
 handlebars.cache=false
 handlebars.expose-session-attributes=true
+
+
